+++ CHANGES.md
@@ -4,8 +4,13 @@
 
 * Make serde an optional dependency for `geo`, rename feature to `use-serde`
   * <https://github.com/georust/rust-geo/pull/209>
+* Use the `proj` crate, rename feature to `use-proj`
+  * <https://github.com/georust/rust-geo/pull/214>
+* Return unboxed iterators from `LineString::lines`, `Winding::points_cw`, and `Winding::points_ccw`
+  * <https://github.com/georust/rust-geo/pull/218>
 * Fix compilation errors when using the `proj` feature
   * <https://github.com/georust/rust-geo/commit/0924f3179c95bfffb847562ee91675d7aa8454f5>
+* Add `Polygon`-`Polygon` and `LineString`-`LineString` distance
 
 ## geo 0.8.3